cleanup organization and continue boilerplate
