+++ Makefile
@@ -1,14 +1,25 @@
+#
+# Globals
+#
 SRC_DIR := ./src
 OBJ_DIR := ./obj
 BIN_DIR := .
 INCLUDE_DIR := ./src #eventually put headers in include/, in src for intellisense
-TEST_DIR := ./tests
 MAIN_NAME := main
+SRC_SUB_DIRS := $(shell find $(SRC_DIR) -type d)
+SQLITE_DIR = ./src/sqlite
 
+
+# Determine key files/dirs
 SRC = $(wildcard $(SRC_DIR)/*.c $(SRC_DIR)/**/*.c)
 OBJ = $(SRC:$(SRC_DIR)/%.c=$(OBJ_DIR)/%.o)
 TARGET = $(BIN_DIR)/$(MAIN_NAME)
 
+OBJ_SUB_DIRS := $(filter-out $(SRC_DIR), $(SRC_SUB_DIRS))
+OBJ_SUB_DIRS := $(shell basename -a $(OBJ_SUB_DIRS))
+OBJ_SUB_DIRS_NOSQLITE := $(filter-out $(shell basename $(SQLITE_DIR)), $(OBJ_SUB_DIRS))
+
+# Compilation and linking parameters
 CC := gcc
 CPPFLAGS := -I$(INCLUDE_DIR) -MMD -MP #c preprocessor
 CFLAGS := -Wformat -Werror=format-security -Wdate-time -D_FORTIFY_SOURCE=2 \
@@ -16,7 +27,8 @@ CFLAGS := -Wformat -Werror=format-security -Wdate-time -D_FORTIFY_SOURCE=2 \
 LDFLAGS := -Llib #linker flag
 LDLIBS := -lm -ldl -lpthread
 
-.PHONY: default all clean test
+# Rules
+.PHONY: default all clean
 .PRECIOUS: $(TARGET) $(OBJ)
 
 default: $(TARGET)
@@ -27,29 +39,35 @@ $(TARGET): $(OBJ)
 	$(CC) $^ $(LDFLAGS) $(LDLIBS) -o $@
 
 # Compilation
-$(OBJ_DIR)/%.o: $(SRC_DIR)/%.c | $(OBJ_DIR)
+$(OBJ_DIR)/%.o: $(SRC_DIR)/%.c | $(OBJ_DIR) $(OBJ_SUB_DIRS)
 	$(CC) $(CPPFLAGS) $(CFLAGS) -c $< -o $@
 
 
-# Other
-$(OBJ_DIR)/%.o: | $(@D)
-	mkdir -p $(@D)
+# Directory setup
+#$(OBJ_DIR)/%.o: | $(@D)
+#	mkdir -p $(@D)
 
+## make ./bin and ./obj
 $(BIN_DIR) $(OBJ_DIR):
 	mkdir -p $@
+
+## make src sub-dirs in obj dir
+$(OBJ_SUB_DIRS):
+	mkdir -p $(OBJ_DIR)/$@
+
+
+
 
 clean:
 	rm -f  $(OBJ_DIR)/*.o
 	rm -f  $(OBJ_DIR)/*.d
-	rm -rf $(TARGET)
-	rm -rf $(BIN_DIR)/test
+	rm -f $(TARGET)
+	cd $(OBJ_DIR) && rm -rf $(OBJ_SUB_DIRS_NOSQLITE)
 
 # sqlite takes longer to build and is not typically modified so no need to
 #   rebuild usually
-clean-sqlite:
-	find $(OBJ_DIR) -type f -name "*.o" -delete
-	find $(OBJ_DIR) -type f -name "*.d" -delete
+
 
 
 # I like this makefile: https://stackoverflow.com/questions/30573481/how-to-write-a-makefile-with-separate-source-and-header-directories
-# Ideal testing setup: https://stackoverflow.com/questions/17896751/makefile-use-multiple-makefiles
+# Ideal testing setup? https://stackoverflow.com/questions/17896751/makefile-use-multiple-makefiles
