+++ src/examples/examples.h
@@ -1,0 +1,6 @@
+#ifndef __EXAMPLES_H
+#define __EXAMPLES_H
+
+#include "examples/read_sqlite_table.h"
+
+#endif // __EXAMPLES_H
