+++ src/tests/run_tests.h
@@ -1,0 +1,9 @@
+#ifndef __RUN_TESTS_H
+#define __RUN_TESTS_H
+
+
+int run_tests(void);
+
+
+
+#endif // __RUN_TESTS_H
