+++ working_examples.c
@@ -68,207 +68,6 @@ int main2() {
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-/*query_buff must be freed*/
-void build_sql_query(char **query_buff, const char *format_string, ...) {
-    va_list argptr;
-    va_start(argptr, format_string);
-    // determine length needed for query
-    int req_len = snprintf(NULL, 0, format_string, argptr); 
-    // allocate memory for query and write it
-    *query_buff = chk_malloc((req_len + 1) * sizeof(char));
-    vsnprintf(*query_buff, req_len + 1, format_string, argptr);
-    va_end(argptr);
-}
-
-
-void exec_sqlite_query(sqlite3 *db,
-                       const char *query,
-                       int (*callback)(void*, int, char**, char**),
-                       void *callback_arg) {
-    char *errptr = NULL;
-    int rc;
-    rc = sqlite3_exec(db, query, callback, callback_arg, &errptr);
-    if (rc != SQLITE_OK) {
-        fprintf(stderr, "sqlite3 error: %s\n", errptr);
-        sqlite3_free(errptr);
-        exit(1);
-    }
-}
-
-
-int print_callback(void *print_header, int argc, char **data, char **columns) {
-    int *ph = (int*)print_header;
-    if (*ph) {
-        for (int i = 0; i < argc; i++) {
-            printf("%s\t", columns[i]);
-        }
-        putchar('\n');
-        (*ph) = 0;
-    }
-    for (int i = 0; i < argc; i++) {
-        printf("%s\t", data[i] ? data[i] : "NULL");
-    }
-    putchar('\n');
-    return 0;
-}
-
-
-void print_head(sqlite3 *db, size_t nrows) {
-    if (nrows == 0) {
-        nrows = 5;
-    }
-    // build query
-    char *query;
-    build_sql_query(&query, "SELECT * FROM birthwt LIMIT %zu", nrows);
-    int print_header = 1;
-    exec_sqlite_query(db,
-                      query,
-                      print_callback,
-                      &print_header);
-    chk_free(query);
-}
-
-
-/*
-    simple query to get the number of rows/cols in the table
-*/
-int nrow_callback(void *n, int argc, char **data, char **columns) {
-    *(size_t*)n = strtoul(data[0], NULL, 10);
-    return 0;
-}
-
-size_t db_nrows(sqlite3 *db) {
-    const char *query = "SELECT COUNT(1) FROM birthwt";
-    size_t n = 0;
-    exec_sqlite_query(db, query, nrow_callback, &n);
-    return n;
-}
-
-
-int ncol_callback(void *n, int argc, char **data, char **columns) {
-    *(size_t*)n = argc;
-    return 0;
-}
-
-size_t db_ncols(sqlite3 *db) {
-    const char *query = "SELECT * FROM birthwt LIMIT 1";
-    size_t n = 0;
-    exec_sqlite_query(db, query, ncol_callback, &n);
-    return n;
-}
-
-int col_type_callback(void *coltypes, int argc, char **data, char **columns) {
-    int *ctypes = (int*)coltypes;
-    size_t col_ix = strtoul(data[0], NULL, 10);
-    if (strcmp(data[2], "INTEGER") == 0) {
-        ctypes[col_ix] = INTS_ARR;
-    } else if (strcmp(data[2], "REAL") == 0) {
-        ctypes[col_ix] = REALS_ARR;
-    } else if (strcmp(data[2], "TEXT") == 0) {
-        ctypes[col_ix] = STRINGS_ARR;
-    } else if (strcmp(data[2], "BLOB") == 0) {
-        ctypes[col_ix] = STRINGS_ARR;
-    } else {
-        // TODO
-        ctypes[col_ix] = NULL_ARR;
-        printf("UNKNOWN - not implemented\n");
-    }
-    return 0;
-}
-
-
-ARRP db_coltypes(sqlite3 *db, size_t ncols) {
-    ARRP v = alloc_array(ncols, INTS_ARR);
-    const char *query = "PRAGMA table_info(birthwt)";
-    exec_sqlite_query(db, query, col_type_callback, integer(v));
-    return v;
-}
-
-
-int colnames_callback(void *vec, int argc, char **data, char **columns) {
-    ARRP *v = (ARRP*)vec;
-    for (int i = 0; i < argc; i++) {
-        set_strings_elt(*v, i, columns[i]);
-    }
-    return 0;
-}
-
-ARRP db_colnames(sqlite3 *db, size_t ncols) {
-    ARRP v = alloc_array(ncols, STRINGS_ARR);
-    const char *query = "SELECT * FROM birthwt LIMIT 1";
-    exec_sqlite_query(db, query, colnames_callback, &v);
-    return v;
-}
-
-
-typedef struct sqlite_table {
-    char *name;
-    char *dbpath;
-    size_t nrows;
-    size_t ncols;
-    ARRP colnames;
-    ARRP coltypes;
-    sqlite3 *db;
-    int ini;
-} sqlite_table;
-
-
-sqlite_table tab = {
-    NULL,
-    NULL,
-    0,
-    0,
-    {0},
-    {0},
-    NULL,
-    0
-};
-
-
-void free_sqlite_table(void);
-
-void init_sqlite_table(const char *dbpath,
-                       const char *table) {
-    
-    if (tab.ini) {
-        fprintf(stderr, "Warning: init_sqlite_table: already initialized, freeing...\n");
-        free_sqlite_table();
-    }
-    printf("Connecting to %s for table '%s'...\n\n", dbpath, table);
-    int rc = sqlite3_open(dbpath, &tab.db);
-    if (rc != SQLITE_OK) {
-        fprintf(stderr, "Can't open database: %s\n", sqlite3_errmsg(tab.db));
-        exit(1);
-    }
-    tab.nrows = db_nrows(tab.db);
-    tab.ncols = db_ncols(tab.db);
-    tab.colnames = db_colnames(tab.db, tab.ncols);
-    tab.coltypes = db_coltypes(tab.db, tab.ncols);
-    chk_strcpy(&tab.name, table);
-    chk_strcpy(&tab.dbpath, dbpath);
-    tab.ini = 1;
-}
-
-
-void free_sqlite_table(void) {
-    printf("\nFreeing sqlite_table...\n");
-    free_array(&tab.colnames);
-    free_array(&tab.coltypes);
-    chk_free(tab.name);
-    chk_free(tab.dbpath);
-    printf("\nClosing database connection...\n");
-    if (tab.db) {
-        int rc = sqlite3_close(tab.db);
-        if (rc != SQLITE_OK) {
-            // TODO ?
-            fprintf(stderr, ":( Failed to close database: %s\n", sqlite3_errmsg(tab.db));
-        }
-        tab.db = NULL;
-    } else{
-        printf("Nothing to do :O\n");
-    }
-    tab.ini = 0;
-}
 
 
 size_t which_str(ARRP v, const char *str) {
@@ -1129,35 +928,3 @@ int main4() {
 }
 
 
-
-int main3() {
-    /*
-            SETUP
-    */
-    if (atexit(free_memstack)) {
-        fprintf(stderr, "Failed to register 'free_memstack'\n");
-        return 1;
-    }
-    if (atexit(free_sqlite_table)) {
-        fprintf(stderr, "Failed to register 'free_sqlite_table'\n");
-        return 1;
-    }
-
-    init_sqlite_table("test.db", "birthwt");
-    /*
-            QUERY TABLE INFO
-    */
-    print_head(tab.db, 10);
-    printf("nrows: %zu\n", tab.nrows);
-    printf("ncols: %zu\n", tab.ncols);
-    /*
-            READ DATA INTO VECTORS...
-    */
-    for (size_t i = 0; i < tab.ncols; i++) {
-        printf("%s (%s) \n",
-               strings_elt(tab.colnames, i),
-               arrtype_str(ints_elt(tab.coltypes, i))
-               );
-    }
-    return 0;
-}
